renamed cache directory to core
