+++ src/core/interop.ts
